dashboard/requirements.txt
streamlit
duckdb
python-dotenv